#2 Implemented the basics of the render system
Added a cRender_System class
Added to cGame_Manager.create_system(string ID) so that it can create a
cRender_System instance
Modified cSystem.update() so it prints the id of the registered
entities.
Removed the 'NAME' requirement for cSystem so that all registered
entites are accepted
Changed how n8::nextID works.  Instead of being a public static, it was
changed to be a private static data member with a static public
get_next_id() method.  When invoked, the method increments the next id
by 1 so it is always unique.