#11 Implemented the camera, and sprite world bounding
Added a controllable componenet to determine which entites can be
controlled
Added height and width to the position component
Added a camera system.  The camera system only cares about 2 entities,
the screen and the entity to follow.  These both have unqique
registration methods
Added methods to the movement system to check if a move will take an
entity outside the bounds of the current world