Implemented world entity binding
Modified the movement system and the movement component to allow the
developer to specify whether an entity is bound to the confines of the
world.  If set to true, when an entity reaches the edge of the world
the sill stop moving.  If set to false, they will reach the edge of the
world, continue past it, and be marked for removal by the game system.