Added AlertDialog and InputDialog classes along with a DialogBuilderInterface class to simplify the implementation of the Builder pattern for Dialogs.
Added logic to GUI.ProcessInput to direct keyboard input to any open InputDialogs.  If none are present, the input is passed to the rest of the gui elements.

Refactored logic in InputBox to account for use of new GUIElement State enums.  Also made minor changes to position and size values in InputBox.

Added the drawing of an outline in the Neutral state for a Button